Move default files creation to post-install
